fix: apply live pitch inline in loopMachine::update, not via AudioConnection
RubberBandWrapper is not an AudioStream subclass — cannot wire via AudioConnection.
Instead: allocate wrapper in audio.cpp, apply feedAudio/retrieveAudio on m_input_buffer
in loopMachine::update() after reading input channels. Also fix narrowing cast in DebugState.

+++ apcKey25Transpose.cpp
@@ -77,7 +77,7 @@ apcKey25::DebugState apcKey25::getDebugState() const
     return {
         m_transposeLocked,
         m_transposePitch,
-        m_pitchWheelOffset,
+        (float)m_pitchWheelOffset,
         m_driftTarget,
         m_computedRatio,
         m_liveEngaged,
